create roomSlice
